pyproject.toml
[project]
name = "flakes-e"
version = "0.1.0"
description = "a django web app for searching movies with tmdb"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "cryptography>=45.0.6",
    "django>=5.2.5",
    "django-allauth>=65.11.1",
    "dotenv>=0.9.9",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
    "tzdata>=2025.2",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "djlint>=1.36.4",
]
